+++ R/2_2_textTrainRegression.R
@@ -1,4 +1,4 @@
-
+#library(magrittr)
 #' Function to find the mode
 #' @param x vector with numbers
 #' @return  Mode value
@@ -13,7 +13,7 @@ statisticalMode <- function(x) {
 #'
 #' @param object An rsplit object (from results_nested_resampling tibble)
 #' object = results_nested_resampling$splits[[1]] OR results_nested_resampling$splits[[1]][[1]]
-#' object = results_nested_resampling$inner_resamples[[1]][[1]][[1]]
+#' object = results_nested_resampling$inner_resamples[[5]][[1]][[1]]
 #' @param penalty hyperparameter for ridge regression.
 #' @param mixture hyperparameter for ridge regression.
 #' @param preprocess_PCA threshold for pca; preprocess_PCA = NA
@@ -22,33 +22,40 @@ statisticalMode <- function(x) {
 #' @return  RMSE.
 #' @noRd
 fit_model_rmse <- function(object, model = "regression", eval_measure = "rmse", penalty = 1, mixture = 0,
-                           preprocess_PCA = NA, variable_name_index_pca = NA, Npredictors = NA) {
+                           preprocess_PCA = NA, variable_name_index_pca = NA, first_n_predictors = NA,
+                           preprocess_step_center = TRUE, preprocess_step_scale = TRUE) {
 
 
   data_train <- rsample::analysis(object)   # mdata_train <- rsample::assessment(object)
   data_train <- tibble::as_tibble(data_train)
 
-  # If testing N first predictors help(step_scale) Npredictors = 3
-  if (!is.na(Npredictors)) {
+  # If testing N first predictors help(step_scale) first_n_predictors = 3
+  if (!is.na(first_n_predictors)) {
       #Select y and id
       Nvariable_totals <- length(data_train)
-      variable_names <- colnames(data_train[(Npredictors+1):(Nvariable_totals-2)])
+      variable_names <- colnames(data_train[(first_n_predictors+1):(Nvariable_totals-2)])
   }else {
     variable_names = "id_nr"
       }
 
-  # Recipe for one embedding input summary(xy_recipe) help(all_of)
+  # Recipe for one embedding input summary(xy_recipe) help(all_of) library(tidyverse)
   if (colnames(data_train[1]) == "Dim1") {
     xy_recipe <- data_train %>%
       recipes::recipe(y ~ .) %>%
       recipes::update_role(dplyr::all_of(variable_names), new_role = "Not_predictors") %>%
       recipes::update_role(id_nr, new_role = "id variable") %>%
-      #recipes::update_role(-id_nr, new_role = "predictor") %>%
       recipes::update_role(y, new_role = "outcome") %>%
-      recipes::step_naomit(Dim1, skip = TRUE) %>%
-      recipes::step_center(recipes::all_predictors()) %>%
-      recipes::step_scale(recipes::all_predictors()) %>%
+      recipes::step_naomit(Dim1, skip = TRUE) #%>%
+
+  if (preprocess_step_center) {
+        xy_recipe <- recipes::step_center(xy_recipe, recipes::all_predictors())
+  }
+  if (preprocess_step_scale) {
+      xy_recipe <- recipes::step_scale(xy_recipe, recipes::all_predictors())
+  }
+
       # If preprocess_PCA is not NULL add PCA step with number of component of % of variance to retain specification
+      xy_recipe <- xy_recipe %>%
       {
         if (!is.na(preprocess_PCA)) {
           if (preprocess_PCA >= 1) {
@@ -61,13 +68,11 @@ fit_model_rmse <- function(object, model = "regression", eval_measure = "rmse", 
         } else {
           .
         }
-      } %>%
-      recipes::prep()
-
+      } #%>%
+      #recipes::prep()
+    xy_recipe_prep <- recipes::prep(xy_recipe)
     # Recipe for multiple word embedding input (with possibility of separate PCAs)
   } else {
-    #V1 <- colnames(data[1])
-    #V1 <- comment(model)
 
     xy_recipe <- data_train %>%
       recipes::recipe(y ~ .) %>%
@@ -75,10 +80,17 @@ fit_model_rmse <- function(object, model = "regression", eval_measure = "rmse", 
       recipes::update_role(id_nr, new_role = "id variable") %>%
       recipes::update_role(-id_nr, new_role = "predictor") %>%
       recipes::update_role(y, new_role = "outcome") %>%
-      recipes::step_naomit(recipes::all_predictors(), skip = TRUE) %>%
-      recipes::step_center(recipes::all_predictors()) %>%
-      recipes::step_scale(recipes::all_predictors())
+      recipes::step_naomit(recipes::all_predictors(), skip = TRUE) #%>%
 
+    if (preprocess_step_center) {
+        xy_recipe <- recipes::step_center(xy_recipe, recipes::all_predictors())
+    }
+    if (preprocess_step_scale) {
+      xy_recipe <- recipes::step_scale(xy_recipe, recipes::all_predictors())
+    }
+
+    # If preprocess_PCA is not NULL add PCA step with number of component of % of variance to retain specification
+    #xy_recipe <- xy_recipe %>%
     # Adding a PCA in each loop; first selecting all variables starting with i="Dim_we1"; and then "Dim_we2" etc
     if (!is.na(preprocess_PCA)) {
       if (preprocess_PCA >= 1) {
@@ -98,28 +110,36 @@ fit_model_rmse <- function(object, model = "regression", eval_measure = "rmse", 
         }
       }
     }
-    xy_recipe <- recipes::prep(xy_recipe)
+    xy_recipe_prep <- recipes::prep(xy_recipe)
   }
 
   # To load the prepared training data into a variable juice() is used.
   # It extracts the data from the xy_recipe object.
-  xy_training <- recipes::juice(xy_recipe)
-
+  nr_predictors <- recipes::juice(xy_recipe_prep)
+  nr_predictors <- length(nr_predictors)
   # Ridge and/or Lasso
-  if (length(xy_training) > 3) {
+  if (nr_predictors > 3) {
     # Create and fit model
-    mod <-
+    mod_spec <-
       {
         if (model == "regression") {
           parsnip::linear_reg(penalty = penalty, mixture = mixture)
         } else if (model == "logistic") parsnip::logistic_reg(mode = "classification", penalty = penalty, mixture = mixture)
       } %>%
-      parsnip::set_engine("glmnet") %>%
-      parsnip::fit(y ~ ., data = xy_training)
+      parsnip::set_engine("glmnet") #%>%
+      #parsnip::fit(y ~ ., data = xy_training)
+
+    # Create Workflow (to know variable roles from recipes) help(workflow)
+    wf <- workflows::workflow() %>%
+      workflows::add_model(mod_spec) %>%
+      workflows::add_recipe(xy_recipe)
+
+    # Fit model
+    mod <-  parsnip::fit(wf, data = data_train)
 
     # Standard regression
-  } else if (length(xy_training) == 3) {
-    mod <-
+  } else if (nr_predictors == 3) {
+    mod_spec <-
       {
         if (model == "regression") {
           parsnip::linear_reg(mode = "regression") %>%
@@ -128,13 +148,20 @@ fit_model_rmse <- function(object, model = "regression", eval_measure = "rmse", 
           parsnip::logistic_reg(mode = "classification") %>%
             parsnip::set_engine("glm")
         }
-      } %>%
-      parsnip::fit(y ~ ., data = xy_training)
+      } #%>%
+      #parsnip::fit(y ~ ., data = xy_training)
+
+    # Create Workflow (to know variable roles from recipes)
+    wf <- workflows::workflow() %>%
+      workflows::add_model(mod_spec) %>%
+      workflows::add_recipe(xy_recipe)
+
+    # Fit model
+    mod <-  parsnip::fit(wf, data = data_train)
   }
+
   # Prepare the test data according to the recipe
-  xy_testing <- xy_recipe %>%
-    recipes::bake(rsample::assessment(object))
-  # look at xy_testing: glimpse(xy_testing)
+  xy_testing <- rsample::assessment(object)
 
   if (model == "regression") {
     # Apply model on new data; penalty
@@ -142,7 +169,7 @@ fit_model_rmse <- function(object, model = "regression", eval_measure = "rmse", 
       stats::predict(mod, xy_testing %>% dplyr::select(-y)) %>%
       dplyr::bind_cols(rsample::assessment(object) %>% dplyr::select(y, id_nr))
 
-    # Get RMSE; eval_measure = "rmse"
+    # Get RMSE; eval_measure = "rmse" library(tidyverse)
     eval_result <- select_eval_measure_val(eval_measure, holdout_pred = holdout_pred, truth = y, estimate = .pred)$.estimate
     # Sort output of RMSE, predictions and truth (observed y)
     output <- list(list(eval_result), list(holdout_pred$.pred), list(holdout_pred$y), list(preprocess_PCA), list(holdout_pred$id_nr))
@@ -202,7 +229,9 @@ fit_model_rmse_wrapper <- function(penalty = penalty,
                                    eval_measure,
                                    preprocess_PCA = preprocess_PCA,
                                    variable_name_index_pca = variable_name_index_pca,
-                                   Npredictors = Npredictors) {
+                                   first_n_predictors = first_n_predictors,
+                                   preprocess_step_center = preprocess_step_center,
+                                   preprocess_step_scale = preprocess_step_scale) {
   fit_model_rmse(object,
     model,
     eval_measure,
@@ -210,7 +239,9 @@ fit_model_rmse_wrapper <- function(penalty = penalty,
     mixture,
     preprocess_PCA = preprocess_PCA,
     variable_name_index_pca = variable_name_index_pca,
-    Npredictors = Npredictors
+    first_n_predictors = first_n_predictors,
+    preprocess_step_center = preprocess_step_center,
+    preprocess_step_scale = preprocess_step_scale
   )
 }
 
@@ -232,7 +263,9 @@ tune_over_cost <- function(object,
                            mixture,
                            preprocess_PCA = preprocess_PCA,
                            variable_name_index_pca = variable_name_index_pca,
-                           Npredictors = Npredictors) {
+                           first_n_predictors = first_n_predictors,
+                           preprocess_step_center = preprocess_step_center,
+                           preprocess_step_scale = preprocess_step_scale) {
 
   # Number of components or percent of variance to attain; min_halving; preprocess_PCA = NULL
   if (!is.na(preprocess_PCA[1])) {
@@ -255,8 +288,8 @@ tune_over_cost <- function(object,
   }
 
   ## Sequence to select dimensions from the semreps. SM-article state: "Adding 1, then multiplying by 1.3 and finally rounding to the nearest integer (e.g., 1, 3, 5, 8, where the next number of dimensions to be tested are the first 12; in other words ([8 +􏰄 1*] 􏱡 1.3)
-  if(!is.na(Npredictors)){
-    stop = Npredictors
+  if(!is.na(first_n_predictors)){
+    stop = first_n_predictors
    new_num = 1
    selection_vector = 1
    while(new_num < stop) {
@@ -264,9 +297,9 @@ tune_over_cost <- function(object,
      selection_vector  <-  c(selection_vector, new_num)
    }
    #Changing the last number to the maximum number of dimensions
-   selection_vector[length(selection_vector)] <- Npredictors
-   Npredictors <- selection_vector
-   Npredictors
+   selection_vector[length(selection_vector)] <- first_n_predictors
+   first_n_predictors <- selection_vector
+   first_n_predictors
   }
 
 
@@ -274,7 +307,7 @@ tune_over_cost <- function(object,
     penalty = penalty,
     mixture = mixture,
     preprocess_PCA = preprocess_PCA_value,
-    Npredictors = Npredictors
+    first_n_predictors = first_n_predictors
   )
 
   # Test models with the different hyperparameters for the inner samples
@@ -282,13 +315,15 @@ tune_over_cost <- function(object,
     grid_inner$penalty,
     grid_inner$mixture,
     grid_inner$preprocess_PCA,
-    grid_inner$Npredictors
+    grid_inner$first_n_predictors
   ),
   fit_model_rmse_wrapper,
   object = object,
   model = model,
   eval_measure = eval_measure,
-  variable_name_index_pca = variable_name_index_pca
+  variable_name_index_pca = variable_name_index_pca,
+  preprocess_step_center = preprocess_step_center,
+  preprocess_step_scale = preprocess_step_scale
   )
 
   # Sort the output to separate the rmse, predictions and truth
@@ -326,7 +361,9 @@ summarize_tune_results <- function(object,
                                    mixture,
                                    preprocess_PCA = preprocess_PCA,
                                    variable_name_index_pca = variable_name_index_pca,
-                                   Npredictors = Npredictors) {
+                                   first_n_predictors = first_n_predictors,
+                                   preprocess_step_center = preprocess_step_center,
+                                   preprocess_step_scale = preprocess_step_scale) {
 
   # Return row-bound tibble containing the INNER results
   purrr::map_df(
@@ -338,36 +375,36 @@ summarize_tune_results <- function(object,
     variable_name_index_pca = variable_name_index_pca,
     model = model,
     eval_measure = eval_measure,
-    Npredictors = Npredictors
+    first_n_predictors = first_n_predictors,
+    preprocess_step_center = preprocess_step_center,
+    preprocess_step_scale = preprocess_step_scale
   )
 }
 
 
 
-
-#x = wordembeddings4$harmonytext
-#y = Language_based_assessment_data_8$hilstotal
-#cv_method = "cv_folds"
+#x <- harmony_we$x
+#y <- data1_csv$HILStot
+#cv_method = "validation_split"
 #outside_folds = 10
 #outside_strata_y = "y"
 #outside_breaks = 4
-#inside_folds = 10
+#inside_folds = 3/4
 #inside_strata_y = "y"
 #inside_breaks = 4
 #model = "regression"
-#eval_measure = "cor_test"
+#eval_measure = "default"
+#preprocess_step_center = TRUE
+#preprocess_step_scale = TRUE
 #preprocess_PCA = NA
-#penalty = 1 #0^seq(-16, 16)
+#penalty = 10^seq(-16, 16)
 #mixture = c(0)
-#Npredictors = NA
+#first_n_predictors = NA
 #method_cor = "pearson"
 #model_description = "Consider writing a description of your model here"
 #multi_cores = "multi_cores_sys_default"
 #save_output = "all"
 #seed = 2020
-
-
-
 
 # devtools::document()
 #' Train word embeddings to a numeric variable.
@@ -376,9 +413,9 @@ summarize_tune_results <- function(object,
 #' @param y Numeric variable to predict.
 #' @param model Type of model. Default is "regression"; see also "logistic" for classification.
 #' @param cv_method Cross-validation method to use within a pipeline of nested outer and inner loops of folds (see nested_cv in rsample).
-#' Default is "cv_folds", which uses rsample::vfold_cv to achieve n-folds in both the outer and inner loops; whereas "validation_split" is
-#' using rsample::validation_split in the inner loop to achieve a development and assessment set (note that for validation_split
-#' the inside_folds should be a proportion, e.g., inside_folds = 3/4).
+#' Default is using cv_fols in the outside folds and "validation_split" using rsample::validation_split in the inner loop to achieve a development and assessment
+#' set (note that for validation_split the inside_folds should be a proportion, e.g., inside_folds = 3/4); whereas "cv_folds" uses rsample::vfold_cv to achieve n-folds
+#' in both the outer and inner loops.
 #' @param outside_folds Number of folds for the outer folds (default = 10).
 #' @param outside_strata_y Variable to stratify according (default y; can set to NULL).
 #' @param outside_breaks The number of bins wanted to stratify a numeric stratification variable in the outer cross-validation loop.
@@ -389,6 +426,8 @@ summarize_tune_results <- function(object,
 #' @param eval_measure Type of evaluative measure to select models from. Default = "rmse" for regression and "bal_accuracy"
 #' for logistic. For regression use "rsq" or "rmse"; and for classification use "accuracy", "bal_accuracy", "sens", "spec", "precision", "kappa", "f_measure",
 #' or "roc_auc",(for more details see the yardstick package).
+#' @param preprocess_step_center normalizes dimensions to have a mean of zero; default is set to TRUE. For more info see (step_center in recipes).
+#' @param preprocess_step_scale  normalize dimensions to have a standard deviation of one. For more info see (step_scale in recipes).
 #' @param preprocess_PCA Pre-processing threshold for PCA (to skip this step set it to NA).
 #' Can select amount of variance to retain (e.g., .90 or as a grid c(0.80, 0.90)); or
 #' number of components to select (e.g., 10). Default is "min_halving", which is a function
@@ -397,7 +436,7 @@ summarize_tune_results <- function(object,
 #' preprocess_PCA = round(max(min(number_features/2), number_participants/2), min(50, number_features))).
 #' @param penalty hyper parameter that is tuned
 #' @param mixture hyper parameter that is tuned default = 0 (hence a pure ridge regression).
-#' @param Npredictors by default this setting is turned off (i.e., NA). To use this method, set it to the highest number of predictors you want to test.
+#' @param first_n_predictors by default this setting is turned off (i.e., NA). To use this method, set it to the highest number of predictors you want to test.
 #' Then the X first dimensions are used in training, using a sequence from Kjell et al., 2019 paper in Psychological Methods.
 #' Adding 1, then multiplying by 1.3 and finally rounding to the nearest integer (e.g., 1, 3, 5, 8).
 #' This option is currently only possible for one embedding at the time.
@@ -435,19 +474,21 @@ summarize_tune_results <- function(object,
 #' @export
 textTrainRegression <- function(x,
                                 y,
-                                cv_method = "cv_folds",
+                                cv_method = "validation_split",
                                 outside_folds = 10,
                                 outside_strata_y = "y",
                                 outside_breaks = 4,
-                                inside_folds = 10,
+                                inside_folds = 3/4,
                                 inside_strata_y = "y",
                                 inside_breaks = 4,
                                 model = "regression",
                                 eval_measure = "default",
+                                preprocess_step_center = TRUE,
+                                preprocess_step_scale = TRUE,
                                 preprocess_PCA = NA,
                                 penalty = 10^seq(-16, 16),
                                 mixture = c(0),
-                                Npredictors = NA,
+                                first_n_predictors = NA,
                                 method_cor = "pearson",
                                 model_description = "Consider writing a description of your model here",
                                 multi_cores = "multi_cores_sys_default",
@@ -530,7 +571,7 @@ textTrainRegression <- function(x,
   xy <- cbind(x2, y)
   xy <- tibble::as_tibble(xy)
   xy$id_nr <- c(seq_len(nrow(xy)))
-
+  xy <- tibble::as_tibble(xy[stats::complete.cases(xy), ])
 
   # Cross-Validation help(nested_cv) help(vfold_cv) help(validation_split) inside_folds = 3/4
   if(cv_method == "cv_folds") {
@@ -592,7 +633,9 @@ textTrainRegression <- function(x,
       mixture = mixture,
       preprocess_PCA = preprocess_PCA,
       variable_name_index_pca = variable_name_index_pca,
-      Npredictors = Npredictors
+      first_n_predictors = first_n_predictors,
+      preprocess_step_center = preprocess_step_center,
+      preprocess_step_scale = preprocess_step_scale
     )
   } else if (multi_cores_use == TRUE) {
     # The multisession plan uses the local cores to process the inner resampling loop.
@@ -609,16 +652,18 @@ textTrainRegression <- function(x,
       mixture = mixture,
       preprocess_PCA = preprocess_PCA,
       variable_name_index_pca = variable_name_index_pca,
-      Npredictors = Npredictors
+      first_n_predictors = first_n_predictors,
+      preprocess_step_center = preprocess_step_center,
+      preprocess_step_scale = preprocess_step_scale
     )
   }
 
 
   # Function to get the lowest eval_measure_val
   if(eval_measure %in% c("accuracy" , "bal_accuracy", "sens", "spec", "precision", "kappa", "f_measure", "roc_auc", "rsq", "cor_test")) {
-    bestParameters <- function(dat) dat[which.max(dat$eval_measure), ]
+    bestParameters <- function(dat) dat[which.max(dat$eval_result), ]
   } else if (eval_measure == "rmse"){
-    bestParameters <- function(dat) dat[which.min(dat$eval_measure), ]
+    bestParameters <- function(dat) dat[which.min(dat$eval_result), ]
   }
 
   # Determine the best parameter estimate from each INNER sample to be used
@@ -626,7 +671,7 @@ textTrainRegression <- function(x,
   hyper_parameter_vals <-
     tuning_results %>%
     purrr::map_df(bestParameters) %>%
-    dplyr::select(c(penalty, mixture, preprocess_PCA, Npredictors))
+    dplyr::select(c(penalty, mixture, preprocess_PCA, first_n_predictors))
 
   # Bind best results
   results_split_parameter <-
@@ -638,16 +683,18 @@ textTrainRegression <- function(x,
   # fit_model_rmse(results_split_parameter$splits[[1]])
   results_outer <- purrr::pmap(
     list(
-      object = results_nested_resampling$splits, # SHOULD WE HERE ONLY USE/PREDICT THE LEFT-OUT PART; not re-running it?
+      object = results_nested_resampling$splits,
       penalty = results_split_parameter$penalty,
       mixture = results_split_parameter$mixture,
       preprocess_PCA = results_split_parameter$preprocess_PCA,
-      Npredictors = results_split_parameter$Npredictors,
+      first_n_predictors = results_split_parameter$first_n_predictors,
       variable_name_index_pca = list(variable_name_index_pca),
       model = model,
-      eval_measure = eval_measure
+      eval_measure = eval_measure,
+      preprocess_step_center = preprocess_step_center,
+      preprocess_step_scale = preprocess_step_scale
     ),
-    fit_model_rmse     # Consider being able to select according to correlations here, rather than rmse ?
+    fit_model_rmse
   )
 
   # Separate RMSE, predictions and observed y
@@ -682,31 +729,39 @@ textTrainRegression <- function(x,
 
   ##### Construct final model to be saved and applied on other data  ########
   ############################################################################
-  xy_short <- xy
+  xy_all <- xy
 
   ######### One word embedding as input
-  if (colnames(xy_short[1]) == "Dim1") {
+  if (colnames(xy_all[1]) == "Dim1") {
 
-    # If testing N first predictors help(step_scale) Npredictors = 3
-    if (!is.na(Npredictors)) {
+    # If testing N first predictors help(step_scale) first_n_predictors = 3
+    if (!is.na(first_n_predictors)) {
       #Select y and id
-      Nvariable_totals <- length(xy_short)
-      variable_names <- colnames(xy_short[(Npredictors+1):(Nvariable_totals-2)])
+      Nvariable_totals <- length(xy_all)
+      variable_names <- colnames(xy_all[(first_n_predictors+1):(Nvariable_totals-2)])
     }else {
       variable_names = "id_nr"
     }
 
     # [0,] is added to just get the col names (and avoid saving all the data with the receipt)
     final_recipe <- # xy %>%
-      recipes::recipe(y ~ ., xy_short[0, ]) %>%
+      recipes::recipe(y ~ ., xy_all[0, ]) %>%
       recipes::update_role(all_of(variable_names), new_role = "Not_predictors") %>%
       recipes::update_role(id_nr, new_role = "id variable") %>%
      # recipes::update_role(-id_nr, new_role = "predictor") %>%
       recipes::update_role(y, new_role = "outcome") %>%
       # recipes::step_BoxCox(all_predictors()) %>%
-      recipes::step_naomit(Dim1, skip = TRUE) %>%
-      recipes::step_center(recipes::all_predictors()) %>%
-      recipes::step_scale(recipes::all_predictors()) %>%
+      recipes::step_naomit(Dim1, skip = TRUE) #%>%
+
+      if (preprocess_step_center) {
+        final_recipe <- recipes::step_center(final_recipe, recipes::all_predictors())
+      }
+    if (preprocess_step_scale) {
+      final_recipe <- recipes::step_scale(final_recipe, recipes::all_predictors())
+    }
+    # If preprocess_PCA is not NULL add PCA step with number of component of % of variance to retain specification
+    final_recipe <- final_recipe %>%
+
       {
         if (!is.na(preprocess_PCA[1])) {
           if (preprocess_PCA[1] >= 1) {
@@ -723,15 +778,20 @@ textTrainRegression <- function(x,
 
     ######### More than one word embeddings as input help(all_of)
   } else {
-    #V1 <- colnames(xy_short[1])
+    #V1 <- colnames(xy_all[1])
 
-    final_recipe <- recipes::recipe(y ~ ., xy_short[0, ]) %>%
+    final_recipe <- recipes::recipe(y ~ ., xy_all[0, ]) %>%
       recipes::update_role(id_nr, new_role = "id variable") %>%
       recipes::update_role(-id_nr, new_role = "predictor") %>%
       recipes::update_role(y, new_role = "outcome") %>%
-      recipes::step_naomit(recipes::all_predictors(), skip = TRUE) %>%
-      recipes::step_center(recipes::all_predictors()) %>%
-      recipes::step_scale(recipes::all_predictors())
+      recipes::step_naomit(recipes::all_predictors(), skip = TRUE) #%>%
+
+    if (preprocess_step_center) {
+      final_recipe <- recipes::step_center(final_recipe, recipes::all_predictors())
+    }
+    if (preprocess_step_scale) {
+      final_recipe <- recipes::step_scale(final_recipe, recipes::all_predictors())
+    }
 
     # Adding a PCA in each loop; first selecting all variables starting with i="Dim_we1"; and then "Dim_we2" etc
     if (!is.na(preprocess_PCA)) {
@@ -760,52 +820,61 @@ textTrainRegression <- function(x,
   # Creating recipe in another environment to avoid saving unnecessarily large parts of the environment
   # when saving the object to rda, rds or Rdata.
   # http://r.789695.n4.nabble.com/Model-object-when-generated-in-a-function-saves-entire-environment-when-saved-td4723192.html
-  recipe_save_small_size <- function(final_recipe, xy_short){
+  recipe_save_small_size <- function(final_recipe, xy_all){
 
     env_final_recipe <- new.env(parent = globalenv())
-    env_final_recipe$xy_short <- xy_short
+    env_final_recipe$xy_all <- xy_all
     env_final_recipe$final_recipe <- final_recipe
 
     with(env_final_recipe, preprocessing_recipe_save <- suppressWarnings(recipes::prep(final_recipe,
-                                                                                       xy_short,
+                                                                                       xy_all,
                                                                                        retain = FALSE)))
   }
 
-  preprocessing_recipe_save <- recipe_save_small_size(final_recipe = final_recipe, xy_short = xy_short)
+  preprocessing_recipe_save <- recipe_save_small_size(final_recipe = final_recipe, xy_all = xy_all)
 
-
-  preprocessing_recipe_use <- recipes::prep(final_recipe, xy_short)
+  #Check number of predictors (to later decide standard or multiple regression)
   # To load the prepared training data into a variable juice() is used.
   # It extracts the data from the xy_recipe object.
-  xy_final <- recipes::juice(preprocessing_recipe_use)
-
+  preprocessing_recipe_prep <- recipes::prep(final_recipe, xy_all)
+  nr_predictors <- recipes::juice(preprocessing_recipe_prep)
+  nr_predictors <- length(nr_predictors)
 
   ####### NEW ENVIRONMENT
-  model_save_small_size <- function(xy_final, xy_short, results_split_parameter, model){
+  model_save_small_size <- function(xy_all, final_recipe, results_split_parameter, model, nr_predictors){
     env_final_model <- new.env(parent = globalenv())
-    env_final_model$xy_final <- xy_final
-    env_final_model$xy_short <- xy_short
+    env_final_model$xy_all       <- xy_all
+    env_final_model$final_recipe <- final_recipe
     env_final_model$penalty_mode <- statisticalMode(results_split_parameter$penalty)
     env_final_model$mixture_mode <- statisticalMode(results_split_parameter$mixture)
-    env_final_model$model <- model
+    env_final_model$model        <- model
+    env_final_model$nr_predictors        <- nr_predictors
     env_final_model$statisticalMode <- statisticalMode
     env_final_model$`%>%`  <-  `%>%`
 
     with(env_final_model,
-  if (length(xy_final) > 3) {
+  if (nr_predictors > 3) {
     # Create and fit model; penalty=NULL mixture = NULL
-    final_predictive_model <-
+    final_predictive_model_spec <-
       {
         if (model == "regression") {
           parsnip::linear_reg(penalty = penalty_mode, mixture = mixture_mode)
         } else if (model == "logistic") parsnip::logistic_reg(mode = "classification", penalty = penalty_mode, mixture = mixture_mode)
       } %>%
-      parsnip::set_engine("glmnet") %>%
-      parsnip::fit(y ~ ., data = xy_final)
+      parsnip::set_engine("glmnet") #%>%
+      #parsnip::fit(y ~ ., data = xy_final)
+
+      # Create Workflow (to know variable roles from recipes) help(workflow)
+      wf_final <- workflows::workflow() %>%
+        workflows::add_model(final_predictive_model_spec) %>%
+        workflows::add_recipe(final_recipe)
+
+      # Fit model
+      final_predictive_model <-  parsnip::fit(wf_final, data = xy_all)
 
     # Standard regression
-  } else if (length(xy_final) == 3) {
-    final_predictive_model <-
+  } else if (nr_predictors == 3) {
+    final_predictive_model_spec <-
       {
         if (model == "regression") {
           parsnip::linear_reg(mode = "regression") %>%
@@ -814,14 +883,22 @@ textTrainRegression <- function(x,
           parsnip::logistic_reg(mode = "classification") %>%
             parsnip::set_engine("glm")
         }
-      } %>%
-      parsnip::fit(y ~ ., data = xy_final)
+      } #%>%
+      #parsnip::fit(y ~ ., data = xy_all)
+
+    # Create Workflow (to know variable roles from recipes) help(workflow)
+    wf_final <- workflows::workflow() %>%
+      workflows::add_model(final_predictive_model_spec) %>%
+      workflows::add_recipe(final_recipe)
+
+    # Fit model
+    final_predictive_model <-  parsnip::fit(wf_final, data = xy_all)
 
   }
 )
   }
 
-  final_predictive_model <- model_save_small_size(xy_final, xy_short, results_split_parameter, model)
+  final_predictive_model <- model_save_small_size(xy_all, final_recipe, results_split_parameter, model, nr_predictors)
 
 
   ##### NEW ENVIRONMENT END
@@ -838,7 +915,7 @@ textTrainRegression <- function(x,
   penalty_setting <- paste("penalty_setting = ", deparse(penalty))
   mixture_setting <- paste("mixture_setting = ", deparse(mixture))
   preprocess_PCA_setting <- paste("preprocess_PCA_setting = ", deparse(preprocess_PCA))
-  Npredictors_setting <- paste("Npredictors_setting = ", deparse(Npredictors))
+  first_n_predictors_setting <- paste("first_n_predictors_setting = ", deparse(first_n_predictors))
 
   # Saving the final mtry and min_n used for the final model.
   penalty_description <- paste("penalty in final model = ", deparse(statisticalMode(results_split_parameter$penalty)))
@@ -850,9 +927,11 @@ textTrainRegression <- function(x,
   preprocess_PCA_description <- paste("preprocess_PCA in final model = ", deparse(statisticalMode(results_split_parameter$preprocess_PCA)))
   preprocess_PCA_fold_description <- paste("preprocess_PCA in each fold = ", deparse(results_split_parameter$preprocess_PCA))
 
-  Npredictors_description <- paste("Npredictors in final model = ", deparse(statisticalMode(results_split_parameter$Npredictors)))
-  Npredictors_fold_description <- paste("Npredictors in each fold = ", deparse(results_split_parameter$Npredictors))
+  first_n_predictors_description <- paste("first_n_predictors in final model = ", deparse(statisticalMode(results_split_parameter$first_n_predictors)))
+  first_n_predictors_fold_description <- paste("first_n_predictors in each fold = ", deparse(results_split_parameter$first_n_predictors))
 
+  preprocess_step_center <- paste("preprocess_step_center_setting = ", deparse(preprocess_step_center))
+  preprocess_step_scale <- paste("preprocess_step_scale_setting = ", deparse(preprocess_step_scale))
 
   # Getting time and date
   T2_textTrainRegression <- Sys.time()
@@ -880,12 +959,14 @@ textTrainRegression <- function(x,
     mixture_setting,
     mixture_description,
     mixture_fold_description,
+    preprocess_step_center,
+    preprocess_step_scale,
     preprocess_PCA_setting,
     preprocess_PCA_description,
     preprocess_PCA_fold_description,
-    Npredictors_setting,
-    Npredictors_description,
-    Npredictors_fold_description,
+    first_n_predictors_setting,
+    first_n_predictors_description,
+    first_n_predictors_fold_description,
     embedding_description,
     model_description,
     time_date
@@ -972,10 +1053,8 @@ textTrainRegression <- function(x,
   remove(results_split_parameter)
   remove(results_outer)
   remove(outputlist_results_outer)
-  remove(xy_short)
+  remove(xy_all)
   remove(final_recipe)
-  remove(preprocessing_recipe_use)
-  remove(xy_final)
 
   final_results
 }